requirements.txt
psycopg2-binary
pandas
numpy
matplotlib
seaborn
python-dotenv
jupyter
